+++ ContentPatcher/Framework/AggregateContextual.cs
@@ -54,6 +54,29 @@ namespace ContentPatcher.Framework
             return this;
         }
 
+        /// <summary>Remove contextual values from the tracker.</summary>
+        /// <param name="value">The context value to stop tracking. Null values are ignored.</param>
+        public AggregateContextual Remove(IContextual value)
+        {
+            if (value != null && this.ValuesImpl.Remove(value))
+            {
+                this.IsMutable = this.ValuesImpl.Any(p => p.IsMutable);
+                this.IsReady = this.ValuesImpl.All(p => p.IsReady);
+            }
+
+            return this;
+        }
+
+        /// <summary>Remove contextual values from the tracker.</summary>
+        /// <param name="values">The context values to stop tracking. Null values are ignored.</param>
+        public AggregateContextual Remove(IEnumerable<IContextual> values)
+        {
+            foreach (IContextual value in values)
+                this.Remove(value);
+
+            return this;
+        }
+
         /// <summary>Update the instance when the context changes.</summary>
         /// <param name="context">Provides access to contextual tokens.</param>
         /// <returns>Returns whether the instance changed.</returns>